requirements.txt
alembic==1.14.0
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Cython==0.29.37
distro==1.9.0
docker==7.1.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==3.1.0
Flask-Dance==7.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==3.2.0
Mako==1.3.8
mariadb==1.1.11
MarkupSafe==3.0.2
numpy==2.2.0
oauthlib==3.2.2
packaging==24.2
paramiko==3.5.1
psutil==6.1.0
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
pyrsistent==0.20.0
python-dotenv==0.21.1
Cython<3
setuptools
PyYAML==5.3.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.36
termcolor==2.5.0
texttable==1.7.0
typing_extensions==4.12.2
urllib3==2.2.3
URLObject==2.4.3
websocket-client==0.59.0
Werkzeug==3.1.3
